+++ tests/config/config_weather.js
@@ -1,6 +1,7 @@
 export default {
   weather_entity: 'weather.la_chapelle_sur_erdre',
   weather_daily_forecast_limit: 10,
+  chart_first_kind_to_render: 'precipitation',
   alert: {
     entity_id: 'sensor.44_weather_alert',
     state_green: 'Vert',